rideshare_scaffold/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic==2.5.0
pydantic[email]==2.5.0
stripe==7.8.0
python-socketio[asyncio_client]==5.10.0
python-dotenv==1.0.0
PyJWT==2.8.0
python-multipart==0.0.6

# AI/RAG Dependencies
langchain==0.1.0
langchain-openai==0.0.5
openai==1.3.7
markdown==3.5.1
pgvector==0.2.4
